chore: rename hooks file
