+++ plugin-hrm-form/src/states/profile/hooks/index.ts
@@ -14,4 +14,4 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-export * from './useProfileFlags';
+export * from './profileFlags';
